+++ packages/mongo-join-builder/tests/index.test.js
@@ -1,4 +1,4 @@
-const { queryParser, pipelineBuilder, mutationBuilder } = require('../');
+const { queryParser, pipelineBuilder } = require('../');
 const { listAdapter } = require('./utils');
 
 describe('Test main export', () => {
@@ -18,63 +18,22 @@ describe('Test main export', () => {
       ],
     };
     const queryTree = queryParser({ listAdapter, getUID: jest.fn(key => key) }, query);
-
-    const aggregateResponse = [
-      {
-        name: 'foobar',
-        age: 23,
-        posts: [1, 3], // the IDs are stored on the field
-        posts_every_posts: [
-          // this is the join result
-          {
-            id: 1,
-            title: 'hello',
-            tags: [4, 5],
-            tags_some_tags: [
-              {
-                id: 4,
-                name: 'foo',
-              },
-              {
-                id: 5,
-                name: 'foo',
-              },
-            ],
-          },
-          {
-            id: 3,
-            title: 'hello',
-            tags: [6],
-            tags_some_tags: [
-              {
-                id: 6,
-                name: 'foo',
-              },
-            ],
-          },
-        ],
-      },
-    ];
     const pipeline = pipelineBuilder(queryTree);
-    const postQueryMutations = mutationBuilder(queryTree.relationships);
-
-    const aggregate = jest.fn(() => Promise.resolve(aggregateResponse));
-    const result = await aggregate(pipeline).then(postQueryMutations);
     expect(pipeline).toMatchObject([
       {
         $lookup: {
           from: 'posts',
           as: 'posts_every_posts',
-          let: { posts_every_posts_ids: { $ifNull: ['$posts', []] } },
+          let: { tmpVar: { $ifNull: ['$posts', []] } },
           pipeline: [
-            { $match: { $expr: { $in: ['$_id', '$$posts_every_posts_ids'] } } },
+            { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
             {
               $lookup: {
                 from: 'tags',
                 as: 'tags_some_tags',
-                let: { tags_some_tags_ids: { $ifNull: ['$tags', []] } },
+                let: { tmpVar: { $ifNull: ['$tags', []] } },
                 pipeline: [
-                  { $match: { $expr: { $in: ['$_id', '$$tags_some_tags_ids'] } } },
+                  { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
                   { $match: { name: { $eq: 'foo' } } },
                   { $addFields: { id: '$_id' } },
                 ],
@@ -89,6 +48,7 @@ describe('Test main export', () => {
               },
             },
             { $addFields: { id: '$_id' } },
+            { $project: { tags_some_tags: 0 } },
           ],
         },
       },
@@ -106,14 +66,7 @@ describe('Test main export', () => {
         },
       },
       { $addFields: { id: '$_id' } },
-    ]);
-
-    expect(result).toMatchObject([
-      {
-        name: 'foobar',
-        age: 23,
-        posts: [1, 3],
-      },
+      { $project: { posts_every_posts: 0 } },
     ]);
   });
 });