+++ packages/mongo-join-builder/tests/join-builder.test.js
@@ -1,4 +1,4 @@
-const { pipelineBuilder, mutationBuilder } = require('../lib/join-builder');
+const { pipelineBuilder } = require('../lib/join-builder');
 
 describe('join builder', () => {
   test('correctly generates joins for simple queries', () => {
@@ -14,7 +14,8 @@ describe('join builder', () => {
     const pipeline = pipelineBuilder({
       matchTerm: { $and: [{ name: { $eq: 'foobar' } }, { age: { $eq: 23 } }] },
       postJoinPipeline: [],
-      relationships: {},
+      excludeFields: [],
+      relationships: [],
     });
 
     expect(pipeline).toMatchObject([
@@ -37,19 +38,20 @@ describe('join builder', () => {
 
     */
     const pipeline = pipelineBuilder({
-      relationships: {
-        abc123: {
+      relationships: [
+        {
           matchTerm: { name: { $eq: 'Alice' } },
           relationshipInfo: {
-            from: 'user-collection',
+            from: 'users',
             field: 'author',
             many: false,
             uniqueField: 'abc123_author',
           },
+          excludeFields: [],
           postJoinPipeline: [],
-          relationships: {},
+          relationships: [],
         },
-      },
+      ],
       matchTerm: {
         $and: [
           { title: { $eq: 'foobar' } },
@@ -57,17 +59,18 @@ describe('join builder', () => {
           { $expr: { $eq: [{ $size: '$abc123_author' }, 1] } },
         ],
       },
+      excludeFields: [],
       postJoinPipeline: [],
     });
 
     expect(pipeline).toMatchObject([
       {
         $lookup: {
-          from: 'user-collection',
+          from: 'users',
           as: 'abc123_author',
-          let: { abc123_author_id: '$author' },
+          let: { tmpVar: '$author' },
           pipeline: [
-            { $match: { $expr: { $eq: ['$_id', '$$abc123_author_id'] } } },
+            { $match: { $expr: { $eq: ['$_id', '$$tmpVar'] } } },
             { $match: { name: { $eq: 'Alice' } } },
             { $addFields: { id: '$_id' } },
           ],
@@ -83,6 +86,7 @@ describe('join builder', () => {
         },
       },
       { $addFields: { id: '$_id' } },
+      { $project: { abc123_author: 0 } },
     ]);
   });
 
@@ -98,18 +102,19 @@ describe('join builder', () => {
 
     */
     const pipeline = pipelineBuilder({
-      relationships: {
-        abc123: {
+      relationships: [
+        {
           relationshipInfo: {
-            from: 'posts-collection',
+            from: 'posts',
             field: 'posts',
             many: true,
             uniqueField: 'abc123_posts',
           },
           postJoinPipeline: [],
-          relationships: {},
+          excludeFields: [],
+          relationships: [],
         },
-      },
+      ],
       matchTerm: {
         $and: [
           { name: { $eq: 'foobar' } },
@@ -117,17 +122,18 @@ describe('join builder', () => {
           { $expr: { $gt: [{ $size: '$abc123_posts' }, 0] } },
         ],
       },
+      excludeFields: [],
       postJoinPipeline: [],
     });
 
     expect(pipeline).toMatchObject([
       {
         $lookup: {
-          from: 'posts-collection',
+          from: 'posts',
           as: 'abc123_posts',
-          let: { abc123_posts_ids: { $ifNull: ['$posts', []] } },
+          let: { tmpVar: { $ifNull: ['$posts', []] } },
           pipeline: [
-            { $match: { $expr: { $in: ['$_id', '$$abc123_posts_ids'] } } },
+            { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
             { $addFields: { id: '$_id' } },
           ],
         },
@@ -142,6 +148,7 @@ describe('join builder', () => {
         },
       },
       { $addFields: { id: '$_id' } },
+      { $project: { abc123_posts: 0 } },
     ]);
   });
 
@@ -157,18 +164,19 @@ describe('join builder', () => {
 
     */
     const pipeline = pipelineBuilder({
-      relationships: {
-        abc123: {
+      relationships: [
+        {
           relationshipInfo: {
-            from: 'posts-collection',
+            from: 'posts',
             field: 'posts',
             many: true,
             uniqueField: 'abc123_posts',
           },
           postJoinPipeline: [{ $orderBy: 'title' }],
-          relationships: {},
+          relationships: [],
+          excludeFields: [],
         },
-      },
+      ],
       matchTerm: {
         $and: [
           { name: { $eq: 'foobar' } },
@@ -176,17 +184,18 @@ describe('join builder', () => {
           { $expr: { $eq: [{ $size: '$abc123_posts' }, { $size: { $ifNull: ['$posts', []] } }] } },
         ],
       },
+      excludeFields: [],
       postJoinPipeline: [{ $limit: 10 }],
     });
 
     expect(pipeline).toMatchObject([
       {
         $lookup: {
-          from: 'posts-collection',
+          from: 'posts',
           as: 'abc123_posts',
-          let: { abc123_posts_ids: { $ifNull: ['$posts', []] } },
+          let: { tmpVar: { $ifNull: ['$posts', []] } },
           pipeline: [
-            { $match: { $expr: { $in: ['$_id', '$$abc123_posts_ids'] } } },
+            { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
             { $addFields: { id: '$_id' } },
             { $orderBy: 'title' },
           ],
@@ -204,6 +213,7 @@ describe('join builder', () => {
         },
       },
       { $addFields: { id: '$_id' } },
+      { $project: { abc123_posts: 0 } },
       { $limit: 10 },
     ]);
   });
@@ -228,20 +238,21 @@ describe('join builder', () => {
 
     */
     const pipeline = pipelineBuilder({
-      relationships: {
-        abc123: {
+      relationships: [
+        {
           matchTerm: {
             $and: [{ title: { $eq: 'hello' } }, { $expr: { $gt: [{ $size: '$def456_tags' }, 0] } }],
           },
           relationshipInfo: {
-            from: 'posts-collection',
+            from: 'posts',
             field: 'posts',
             many: true,
             uniqueField: 'abc123_posts',
           },
           postJoinPipeline: [],
-          relationships: {
-            def456: {
+          excludeFields: [],
+          relationships: [
+            {
               matchTerm: {
                 $and: [
                   { name: { $eq: 'React' } },
@@ -253,29 +264,31 @@ describe('join builder', () => {
                 ],
               },
               relationshipInfo: {
-                from: 'tags-collection',
+                from: 'tags',
                 field: 'tags',
                 many: true,
                 uniqueField: 'def456_tags',
               },
               postJoinPipeline: [],
-              relationships: {
-                xyz890: {
+              excludeFields: [],
+              relationships: [
+                {
                   matchTerm: { published: { $eq: true } },
                   relationshipInfo: {
-                    from: 'posts-collection',
+                    from: 'posts',
                     field: 'posts',
                     many: true,
                     uniqueField: 'xyz890_posts',
                   },
                   postJoinPipeline: [],
-                  relationships: {},
+                  excludeFields: [],
+                  relationships: [],
                 },
-              },
+              ],
             },
-          },
+          ],
         },
-      },
+      ],
       matchTerm: {
         $and: [
           { name: { $eq: 'foobar' } },
@@ -283,31 +296,32 @@ describe('join builder', () => {
           { $expr: { $gt: [{ $size: '$abc123_posts' }, 0] } },
         ],
       },
+      excludeFields: [],
       postJoinPipeline: [],
     });
 
     expect(pipeline).toMatchObject([
       {
         $lookup: {
-          from: 'posts-collection',
+          from: 'posts',
           as: 'abc123_posts',
-          let: { abc123_posts_ids: { $ifNull: ['$posts', []] } },
+          let: { tmpVar: { $ifNull: ['$posts', []] } },
           pipeline: [
-            { $match: { $expr: { $in: ['$_id', '$$abc123_posts_ids'] } } },
+            { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
             {
               $lookup: {
-                from: 'tags-collection',
+                from: 'tags',
                 as: 'def456_tags',
-                let: { def456_tags_ids: { $ifNull: ['$tags', []] } },
+                let: { tmpVar: { $ifNull: ['$tags', []] } },
                 pipeline: [
-                  { $match: { $expr: { $in: ['$_id', '$$def456_tags_ids'] } } },
+                  { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
                   {
                     $lookup: {
-                      from: 'posts-collection',
+                      from: 'posts',
                       as: 'xyz890_posts',
-                      let: { xyz890_posts_ids: { $ifNull: ['$posts', []] } },
+                      let: { tmpVar: { $ifNull: ['$posts', []] } },
                       pipeline: [
-                        { $match: { $expr: { $in: ['$_id', '$$xyz890_posts_ids'] } } },
+                        { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
                         { $match: { published: { $eq: true } } },
                         { $addFields: { id: '$_id' } },
                       ],
@@ -330,6 +344,7 @@ describe('join builder', () => {
                     },
                   },
                   { $addFields: { id: '$_id' } },
+                  { $project: { xyz890_posts: 0 } },
                 ],
               },
             },
@@ -342,6 +357,7 @@ describe('join builder', () => {
               },
             },
             { $addFields: { id: '$_id' } },
+            { $project: { def456_tags: 0 } },
           ],
         },
       },
@@ -355,6 +371,7 @@ describe('join builder', () => {
         },
       },
       { $addFields: { id: '$_id' } },
+      { $project: { abc123_posts: 0 } },
     ]);
   });
 
@@ -376,8 +393,8 @@ describe('join builder', () => {
     */
 
     const pipeline = pipelineBuilder({
-      relationships: {
-        zip567: {
+      relationships: [
+        {
           matchTerm: {
             $and: [
               { title: { $eq: 'hello' } },
@@ -385,27 +402,29 @@ describe('join builder', () => {
             ],
           },
           relationshipInfo: {
-            from: 'posts-collection',
+            from: 'posts',
             field: 'posts',
             many: true,
             uniqueField: 'zip567_posts',
           },
           postJoinPipeline: [],
-          relationships: {
-            quux987: {
+          excludeFields: [],
+          relationships: [
+            {
               matchTerm: { name: { $eq: 'foo' } },
               relationshipInfo: {
-                from: 'labels-collection',
+                from: 'labels',
                 field: 'labels',
                 many: true,
                 uniqueField: 'quux987_labels',
               },
               postJoinPipeline: [],
-              relationships: {},
+              excludeFields: [],
+              relationships: [],
             },
-          },
+          ],
         },
-      },
+      ],
       matchTerm: {
         $and: [
           { name: { $eq: 'foobar' } },
@@ -417,24 +436,25 @@ describe('join builder', () => {
           },
         ],
       },
+      excludeFields: [],
       postJoinPipeline: [],
     });
 
     expect(pipeline).toMatchObject([
       {
         $lookup: {
-          from: 'posts-collection',
+          from: 'posts',
           as: 'zip567_posts',
-          let: { zip567_posts_ids: { $ifNull: ['$posts', []] } },
+          let: { tmpVar: { $ifNull: ['$posts', []] } },
           pipeline: [
-            { $match: { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } } },
+            { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
             {
               $lookup: {
-                from: 'labels-collection',
+                from: 'labels',
                 as: 'quux987_labels',
-                let: { quux987_labels_ids: { $ifNull: ['$labels', []] } },
+                let: { tmpVar: { $ifNull: ['$labels', []] } },
                 pipeline: [
-                  { $match: { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } } },
+                  { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
                   { $match: { name: { $eq: 'foo' } } },
                   { $addFields: { id: '$_id' } },
                 ],
@@ -449,6 +469,7 @@ describe('join builder', () => {
               },
             },
             { $addFields: { id: '$_id' } },
+            { $project: { quux987_labels: 0 } },
           ],
         },
       },
@@ -466,6 +487,7 @@ describe('join builder', () => {
         },
       },
       { $addFields: { id: '$_id' } },
+      { $project: { zip567_posts: 0 } },
     ]);
   });
 
@@ -487,8 +509,8 @@ describe('join builder', () => {
     */
 
     const pipeline = pipelineBuilder({
-      relationships: {
-        zip567: {
+      relationships: [
+        {
           matchTerm: {
             $or: [
               { title: { $eq: 'hello' } },
@@ -496,27 +518,29 @@ describe('join builder', () => {
             ],
           },
           relationshipInfo: {
-            from: 'posts-collection',
+            from: 'posts',
             field: 'posts',
             many: true,
             uniqueField: 'zip567_posts',
           },
           postJoinPipeline: [],
-          relationships: {
-            quux987: {
+          excludeFields: [],
+          relationships: [
+            {
               matchTerm: { name: { $eq: 'foo' } },
               relationshipInfo: {
-                from: 'labels-collection',
+                from: 'labels',
                 field: 'labels',
                 many: true,
                 uniqueField: 'quux987_labels',
               },
               postJoinPipeline: [],
-              relationships: {},
+              excludeFields: [],
+              relationships: [],
             },
-          },
+          ],
         },
-      },
+      ],
       matchTerm: {
         $or: [
           { name: { $eq: 'foobar' } },
@@ -528,27 +552,26 @@ describe('join builder', () => {
           },
         ],
       },
+      excludeFields: [],
       postJoinPipeline: [],
     });
 
     expect(pipeline).toMatchObject([
       {
         $lookup: {
-          from: 'posts-collection',
+          from: 'posts',
           as: 'zip567_posts',
-          let: { zip567_posts_ids: { $ifNull: ['$posts', []] } },
+          let: { tmpVar: { $ifNull: ['$posts', []] } },
           pipeline: [
-            { $match: { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } } },
+            { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
             {
               $lookup: {
-                from: 'labels-collection',
+                from: 'labels',
                 as: 'quux987_labels',
-                let: { quux987_labels_ids: { $ifNull: ['$labels', []] } },
+                let: { tmpVar: { $ifNull: ['$labels', []] } },
                 pipeline: [
-                  { $match: { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } } },
-                  {
-                    $match: { name: { $eq: 'foo' } },
-                  },
+                  { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
+                  { $match: { name: { $eq: 'foo' } } },
                   { $addFields: { id: '$_id' } },
                 ],
               },
@@ -562,6 +585,7 @@ describe('join builder', () => {
               },
             },
             { $addFields: { id: '$_id' } },
+            { $project: { quux987_labels: 0 } },
           ],
         },
       },
@@ -579,6 +603,7 @@ describe('join builder', () => {
         },
       },
       { $addFields: { id: '$_id' } },
+      { $project: { zip567_posts: 0 } },
     ]);
   });
 
@@ -600,8 +625,8 @@ describe('join builder', () => {
     */
 
     const pipeline = pipelineBuilder({
-      relationships: {
-        zip567: {
+      relationships: [
+        {
           matchTerm: {
             $or: [
               { title: { $eq: 'hello' } },
@@ -610,27 +635,29 @@ describe('join builder', () => {
           },
 
           relationshipInfo: {
-            from: 'posts-collection',
+            from: 'posts',
             field: 'posts',
             many: true,
             uniqueField: 'zip567_posts',
           },
           postJoinPipeline: [],
-          relationships: {
-            quux987: {
+          excludeFields: [],
+          relationships: [
+            {
               matchTerm: { name: { $eq: 'foo' } },
               relationshipInfo: {
-                from: 'labels-collection',
+                from: 'labels',
                 field: 'labels',
                 many: true,
                 uniqueField: 'quux987_labels',
               },
               postJoinPipeline: [],
-              relationships: {},
+              excludeFields: [],
+              relationships: [],
             },
-          },
+          ],
         },
-      },
+      ],
       matchTerm: {
         $and: [
           { name: { $eq: 'foobar' } },
@@ -642,27 +669,26 @@ describe('join builder', () => {
           },
         ],
       },
+      excludeFields: [],
       postJoinPipeline: [],
     });
 
     expect(pipeline).toMatchObject([
       {
         $lookup: {
-          from: 'posts-collection',
+          from: 'posts',
           as: 'zip567_posts',
-          let: { zip567_posts_ids: { $ifNull: ['$posts', []] } },
+          let: { tmpVar: { $ifNull: ['$posts', []] } },
           pipeline: [
-            { $match: { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } } },
+            { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
             {
               $lookup: {
-                from: 'labels-collection',
+                from: 'labels',
                 as: 'quux987_labels',
-                let: { quux987_labels_ids: { $ifNull: ['$labels', []] } },
+                let: { tmpVar: { $ifNull: ['$labels', []] } },
                 pipeline: [
-                  { $match: { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } } },
-                  {
-                    $match: { name: { $eq: 'foo' } },
-                  },
+                  { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
+                  { $match: { name: { $eq: 'foo' } } },
                   { $addFields: { id: '$_id' } },
                 ],
               },
@@ -676,6 +702,7 @@ describe('join builder', () => {
               },
             },
             { $addFields: { id: '$_id' } },
+            { $project: { quux987_labels: 0 } },
           ],
         },
       },
@@ -693,6 +720,7 @@ describe('join builder', () => {
         },
       },
       { $addFields: { id: '$_id' } },
+      { $project: { zip567_posts: 0 } },
     ]);
   });
 
@@ -714,8 +742,8 @@ describe('join builder', () => {
     */
 
     const pipeline = pipelineBuilder({
-      relationships: {
-        zip567: {
+      relationships: [
+        {
           matchTerm: {
             $and: [
               { title: { $eq: 'hello' } },
@@ -723,56 +751,55 @@ describe('join builder', () => {
             ],
           },
           relationshipInfo: {
-            from: 'posts-collection',
+            from: 'posts',
             field: 'posts',
             many: true,
             uniqueField: 'zip567_posts',
           },
           postJoinPipeline: [],
-          relationships: {
-            quux987: {
+          excludeFields: [],
+          relationships: [
+            {
               matchTerm: { name: { $eq: 'foo' } },
               relationshipInfo: {
-                from: 'labels-collection',
+                from: 'labels',
                 field: 'labels',
                 many: true,
                 uniqueField: 'quux987_labels',
               },
               postJoinPipeline: [],
-              relationships: {},
+              excludeFields: [],
+              relationships: [],
             },
-          },
+          ],
         },
-      },
+      ],
       matchTerm: {
         $or: [
           { name: { $eq: 'foobar' } },
           { age: { $eq: 23 } },
-          {
-            $expr: {
-              $eq: [{ $size: '$zip567_posts' }, { $size: { $ifNull: ['$posts', []] } }],
-            },
-          },
+          { $expr: { $eq: [{ $size: '$zip567_posts' }, { $size: { $ifNull: ['$posts', []] } }] } },
         ],
       },
+      excludeFields: [],
       postJoinPipeline: [],
     });
 
     expect(pipeline).toMatchObject([
       {
         $lookup: {
-          from: 'posts-collection',
+          from: 'posts',
           as: 'zip567_posts',
-          let: { zip567_posts_ids: { $ifNull: ['$posts', []] } },
+          let: { tmpVar: { $ifNull: ['$posts', []] } },
           pipeline: [
-            { $match: { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } } },
+            { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
             {
               $lookup: {
-                from: 'labels-collection',
+                from: 'labels',
                 as: 'quux987_labels',
-                let: { quux987_labels_ids: { $ifNull: ['$labels', []] } },
+                let: { tmpVar: { $ifNull: ['$labels', []] } },
                 pipeline: [
-                  { $match: { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } } },
+                  { $match: { $expr: { $in: ['$_id', '$$tmpVar'] } } },
                   { $match: { name: { $eq: 'foo' } } },
                   { $addFields: { id: '$_id' } },
                 ],
@@ -787,6 +814,7 @@ describe('join builder', () => {
               },
             },
             { $addFields: { id: '$_id' } },
+            { $project: { quux987_labels: 0 } },
           ],
         },
       },
@@ -796,85 +824,13 @@ describe('join builder', () => {
             { name: { $eq: 'foobar' } },
             { age: { $eq: 23 } },
             {
-              $expr: {
-                $eq: [{ $size: '$zip567_posts' }, { $size: { $ifNull: ['$posts', []] } }],
-              },
+              $expr: { $eq: [{ $size: '$zip567_posts' }, { $size: { $ifNull: ['$posts', []] } }] },
             },
           ],
         },
       },
       { $addFields: { id: '$_id' } },
+      { $project: { zip567_posts: 0 } },
     ]);
-  });
-
-  test('executes relationship mutators with correct parameters', () => {
-    // TODO - check it's called with these params:
-    /*
-     * From this query:
-
-      {
-        age: 23,
-        posts_every: { title: 'hello' },
-      }
-    */
-
-    const mutationResult = {};
-
-    const postQueryMutations = mutationBuilder({
-      zip567: {
-        matchTerm: { title: { $eq: 'hello' } },
-
-        relationshipInfo: {
-          from: 'posts-collection',
-          field: 'posts',
-          many: true,
-          uniqueField: 'zip567_posts',
-        },
-        postJoinPipeline: [],
-        relationships: {},
-      },
-    });
-
-    /*
-      {
-        age: 23,
-        posts_every: { title: 'hello' },
-      }
-    */
-    const mockResult1 = {
-      age: 23,
-      name: 'foobar',
-      zip567_posts: [
-        {
-          title: 'hello',
-          views: 73,
-        },
-        {
-          title: 'hello',
-          views: 57,
-        },
-      ],
-    };
-
-    const mockResult2 = {
-      age: 23,
-      name: 'quux',
-      zip567_posts: [
-        {
-          title: 'hello',
-          views: 123,
-        },
-        {
-          title: 'hello',
-          views: 1,
-        },
-      ],
-    };
-
-    const mockQueryResult = [mockResult1, mockResult2];
-
-    const mutatedResult = postQueryMutations(mockQueryResult);
-
-    expect(mutatedResult).toMatchObject([mutationResult, mutationResult]);
   });
 });